Re-org tests to isolate native func return logic
