+++ ulox/ulox.core.tests/MutliReturnTests.cs
@@ -64,100 +64,6 @@ print(b);");
 //        }
 
         [Test]
-        public void Run_WhenReturnNothingFromExternalFunctionInMiddleOfMathOpsAndAllLocals_ShouldNotImpactMathOps()
-        {
-            testEngine.MyEngine.Context.Vm.SetGlobal(new HashedString("A"), Value.New(ReturnNothing));
-
-            testEngine.Run(@"
-fun Locals()
-{
-    var a = 1;
-    var b = 2;
-    A();
-
-    a += b + 1;
-
-    print(a);
-}
-
-Locals();");
-
-            Assert.AreEqual("4", testEngine.InterpreterResult);
-        }
-
-        [Test]
-        public void Run_WhenReturnNothing_ShouldNotHaveError()
-        {
-            testEngine.MyEngine.Context.Vm.SetGlobal(new HashedString("A"), Value.New(ReturnNothing));
-
-            testEngine.Run(@"A();");
-
-            Assert.AreEqual("", testEngine.InterpreterResult);
-        }
-
-        [Test]
-        public void Run_WhenReturnNothingFromExternalFunctionInMiddleOfMathOps_ShouldNotImpactMathOps()
-        {
-            testEngine.MyEngine.Context.Vm.SetGlobal(new HashedString("A"), Value.New(ReturnNothing));
-
-            testEngine.Run(@"
-var a = 1;
-var b = 2;
-A();
-
-a += b + 1;
-
-print(a);");
-
-            Assert.AreEqual("4", testEngine.InterpreterResult);
-        }
-
-        private NativeCallResult ReturnNothing(Vm vm, int argc)
-        {
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        [Test]
-        public void Run_WhenReturnFromExternalFunction_ShouldMatchExpected()
-        {
-            testEngine.MyEngine.Context.Vm.SetGlobal(new HashedString("A1"), Value.New(Return1Thing));
-
-            testEngine.Run(@"
-var a = A1();
-
-print(a);");
-
-            Assert.AreEqual("1", testEngine.InterpreterResult);
-        }
-
-        private NativeCallResult Return1Thing(Vm vm, int arg2)
-        {
-            vm.PushReturn(Value.New(1));
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        [Test]
-        public void Run_WhenReturn2FromExternalFunction_ShouldMatchExpected()
-        {
-            testEngine.MyEngine.Context.Vm.SetGlobal(new HashedString("A2"), Value.New(Return2Thing));
-
-            testEngine.Run(@"
-var (a,b) = A2();
-
-print(a);
-print(b);");
-
-            Assert.AreEqual("12", testEngine.InterpreterResult);
-        }
-
-        private NativeCallResult Return2Thing(Vm vm, int arg2)
-        {
-            vm.PushReturn(Value.New(1));
-            vm.PushReturn(Value.New(2));
-            return NativeCallResult.SuccessfulExpression;
-        }
-
-        [Test]
         public void Run_WhenReturn2AndTake2_ShouldMatchExpected()
         {
             testEngine.Run(@"
